kapil/ollama_LLM_models/requirements.txt
<<<<<<< HEAD
python-dotenv
requests
langchain-community
langchain-openai
streamlit
=======
streamlit

>>>>>>> adding local_conversion of part1 code
